Moved to alternating pattern for the front panel input (TrigT1CTP-03-00-76)
	* Moved to alternating pattern for the front panel input
	* TrigT1CTP-03-00-75/6

2015-09-21 Joerg Stelzer
	* Restore some histogramming functionality accidentaly removed
	* TrigT1CTP-03-00-74

2015-09-20 Joerg Stelzer
	* Move histogram booking from start to beginRun
	* TrigT1CTP-03-00-72/73

2015-09-20 Joerg Stelzer
	* Remove zero bias trigger from simulation
	* TrigT1CTP-03-00-71

2015-09-17 Joerg Stelzer
	* Fix CTP simulation
	* TrigT1CTP-03-00-70

2015-08-09 Joerg Stelzer
...
(Long ChangeLog diff - truncated)

+++ Trigger/TrigT1/TrigT1CTP/TrigT1CTP/CTPUtil.h
@@ -49,7 +49,7 @@ namespace LVL1CTP {
     static int getTMMult( unsigned int word, int threshold );     //!< get missing Et multiplicty in legacy mode
 
     //! extract multiplicities using new trigger configuration interface
-    static int getMult( unsigned int word, int startbit, int endbit );
+    static int getMult( uint64_t word, unsigned int startbit, unsigned int endbit );
 
     // Some additional helper functions:
     static unsigned int alignBits( int value, int startbit, int endbit );            //!< align bits
